Refactor the channel settings dialog into a full screen overlay.
The channel settings are now rendered in the same way our regular
settings are. In fact, I'm actually sharing a lot of the same UI
components between the two.

Even though the channel settings only has one page for now, we'll soon
add more: ability to change permissions of channel members, ability to
see banned users and unban them, invitations in case of the private
channels, etc.

Currently I haven't implemented the logic which would prevent user from
navigating to a different channel settings page, or close the channel
settings overlay, in case there are some changes. I wasn't exactly sure
what's the best way to deal with that at the moment.
