+++ client/chat/channel-settings/channel-settings-page.tsx
@@ -1,0 +1,9 @@
+export enum GeneralChannelSettingsPage {
+  General = 'General',
+}
+
+export type ChannelSettingsPage = GeneralChannelSettingsPage
+
+export const ALL_CHANNEL_SETTINGS_PAGES: ReadonlyArray<ChannelSettingsPage> = [
+  ...Object.values(GeneralChannelSettingsPage),
+]
